+++ frontend/tests/colonistData.test.ts
@@ -1,7 +1,10 @@
 import ColonistData from "../src/colonistData";
+import Infrastructure from "../src/infrastructure";
+jest.mock("../src/infrastructure");
 
 describe("ColonistData", () => {
     const colonistData = new ColonistData(1, 0);
+    const mockInfra = new Infrastructure();
 
     const flatTerrain = [[1, 1, 1], [1, 1, 1], [1, 1, 1]];
     // Heights per column: 3, 4, 6 (max three columns shown)
@@ -29,7 +32,7 @@ describe("ColonistData", () => {
         colonistData._isMoving = true;
         colonistData._movementCost = 10;
         colonistData._movementProg = 0;
-        colonistData.checkGoalStatus(flatTerrain, 3);
+        colonistData.checkGoalStatus(flatTerrain, 3, mockInfra);
         expect(colonistData._movementProg).toBe(1);
     })
 
@@ -39,7 +42,7 @@ describe("ColonistData", () => {
         colonistData._x = 1;
         colonistData._movementDest = 0;
         colonistData._isMoving = false;
-        colonistData.checkGoalStatus(flatTerrain, 3);
+        colonistData.checkGoalStatus(flatTerrain, 3, mockInfra);
         expect(colonistData._isMoving).toBe(true);
         expect(colonistData._movementType).toBe("walk");
         expect(colonistData._y).toBe(31);       // y = 31 when the current column has 3 blocks
@@ -51,7 +54,7 @@ describe("ColonistData", () => {
         colonistData._x = 1;                // Start away from the edge
         colonistData._y = 0;                // Start the colonist up in the air to test their gravity as well!
         colonistData._movementDest = 4;     // Destination is to the right - current column = 4, dest = 6
-        colonistData.checkGoalStatus(unevenTerrain, 5);
+        colonistData.checkGoalStatus(unevenTerrain, 5, mockInfra);
         expect(colonistData._isMoving).toBe(true);
         expect(colonistData._y).toBe(30);   // Column is 4 blocks tall
         expect(colonistData._movementType).toBe("big-climb");
@@ -64,7 +67,7 @@ describe("ColonistData", () => {
         colonistData._x = 4;                // Start away from the edge
         colonistData._y = 0;                // Start the colonist up in the air to test their gravity as well!
         colonistData._movementDest = 0;     // Destination is to the right - current column = 4, dest = 6
-        colonistData.checkGoalStatus(unevenTerrain, 5);
+        colonistData.checkGoalStatus(unevenTerrain, 5, mockInfra);
         expect(colonistData._isMoving).toBe(true);
         expect(colonistData._movementType).toBe("small-drop");
         expect(colonistData._facing).toBe("left");