+++ ChangeLog.md
@@ -1594,7 +1594,9 @@ Features Added:
 
 26. Next, create another new field for the colonist for activityTimeElapsed, which will be the amount of minutes that have passed while doing the current activity (if its duration is more than 0).
 
-### 26. Create a new Colonist Data method (unit test first!) that will
+27. Update the Engine's call to the Population class's updateColonists method to pass it a link to the Infrastructure class, so that it can pass that to individual colonists to use its methods.
+
+### 28. Remove P5 from the constructor of the Infrastructure class; replace both uses of it (there are only 2, one for module creation and one for connector creation) by passing the Engine's P5 instance to those method calls rather than using them in the Infra base class's constructor.
 
 ### 99 [Optional] For your inner bureaucrat: re-arrange it so that all of the type definitions are exported from a single file, since the current scheme is getting a bit confusing.